feat(management): 获取家长UID-bug fix

+++ myopia-business/business-api/parent-api/src/main/java/com/wupol/myopia/business/api/parent/domain/dto/ParentUidRequestDTO.java
@@ -25,7 +25,7 @@ public class ParentUidRequestDTO {
     /**
      * timestamp
      */
-    private Integer timestamp;
+    private Long timestamp;
 
     /**
      * url
